// src/models/the-guardian.types.ts
export interface TheGuardianArticle {
  id: string;
  type: string;
  sectionId: string;
  sectionName: string;
  webPublicationDate: string;
  webTitle: string;
  webUrl: string;
  apiUrl: string;
  isHosted: boolean;
  pillarId: string;
  pillarName: string;
}

export interface TheGuardianResponseData {
  status: string;
  userTier: string;
  total: number;
  startIndex: number;
  pageSize: number;
  currentPage: number;
  pages: number;
  orderBy: string;
  results: TheGuardianArticle[];
}

export interface TheGuardianApiResponse {
  response: TheGuardianResponseData;
}